+++ pegasus/src/Pegasus/IndicationService/tests/IndicationService/TestIndicationService.cpp
@@ -67,12 +67,12 @@ void test_setSystemNameInHandlerFilter()
 
     source.set(fullSourcePath);
     PEGASUS_TEST_ASSERT(source != emptyTarget);
-    IndicationService::_setSystemNameInHandlerFilter(source,String::EMPTY);
+    IndicationService::_setSystemNameInHandlerFilterPath(source,String::EMPTY);
     PEGASUS_TEST_ASSERT(source == emptyTarget);
 
     source.set(fullSourcePath);
     PEGASUS_TEST_ASSERT(source != setTarget);
-    IndicationService::_setSystemNameInHandlerFilter(source,"TestSys");
+    IndicationService::_setSystemNameInHandlerFilterPath(source,"TestSys");
     PEGASUS_TEST_ASSERT(source == setTarget);
 }
 
@@ -173,11 +173,145 @@ void test_setSystemNameInHandlerFilterReference()
     PEGASUS_TEST_ASSERT(fullSourcePath == fullSetTargetPath);
 }
 
+CIMObjectPath _buildFilterOrHandlerPath
+    (const CIMName & className,
+     const String & name,
+     const String & host,
+     const CIMNamespaceName & namespaceName = CIMNamespaceName())
+{
+    CIMObjectPath path;
+
+    Array <CIMKeyBinding> keyBindings;
+    keyBindings.append (CIMKeyBinding ("SystemCreationClassName",
+        System::getSystemCreationClassName(), CIMKeyBinding::STRING));
+    keyBindings.append (CIMKeyBinding ("SystemName",
+        System::getFullyQualifiedHostName(), CIMKeyBinding::STRING));
+    keyBindings.append (CIMKeyBinding ("CreationClassName",
+        className.getString(), CIMKeyBinding::STRING));
+    keyBindings.append (CIMKeyBinding ("Name", name, CIMKeyBinding::STRING));
+    path.setClassName (className);
+    path.setKeyBindings (keyBindings);
+    path.setNameSpace (namespaceName);
+    path.setHost (host);
+
+    return path;
+}
+
+CIMObjectPath _buildSubscriptionPath
+    (const String & filterName,
+     const CIMName & handlerClass,
+     const String & handlerName,
+     const String & filterHost,
+     const String & handlerHost,
+     const CIMNamespaceName & filterNS,
+     const CIMNamespaceName & handlerNS)
+{
+    CIMObjectPath filterPath = _buildFilterOrHandlerPath
+        (PEGASUS_CLASSNAME_INDFILTER, filterName, filterHost, filterNS);
+
+    CIMObjectPath handlerPath = _buildFilterOrHandlerPath (handlerClass,
+        handlerName, handlerHost, handlerNS);
+
+    Array<CIMKeyBinding> subscriptionKeyBindings;
+    subscriptionKeyBindings.append (CIMKeyBinding ("Filter",
+        filterPath.toString(), CIMKeyBinding::REFERENCE));
+    subscriptionKeyBindings.append (CIMKeyBinding ("Handler",
+        handlerPath.toString(), CIMKeyBinding::REFERENCE));
+    CIMObjectPath subscriptionPath ("", CIMNamespaceName(),
+        PEGASUS_CLASSNAME_INDSUBSCRIPTION, subscriptionKeyBindings);
+
+    return subscriptionPath;
+}
+
+// Build subscription with empty 
+CIMObjectPath _buildSubscriptionPath
+    (const String & filterName,
+     const CIMName & handlerClass,
+     const String & handlerName)
+{
+    return _buildSubscriptionPath(filterName, handlerClass, handlerName,
+        String::EMPTY, String::EMPTY, CIMNamespaceName(), CIMNamespaceName());
+}
+
+CIMInstance _buildSubscriptionInstance
+    (const CIMObjectPath & filterPath,
+     const CIMName & handlerClass,
+     const CIMObjectPath & handlerPath)
+{
+    CIMInstance subscriptionInstance (PEGASUS_CLASSNAME_INDSUBSCRIPTION);
+
+    CIMObjectPath path = _buildSubscriptionPath ("Filter01",
+        PEGASUS_CLASSNAME_INDHANDLER_CIMXML, "Handler01");
+    subscriptionInstance.setPath (path);
+
+    subscriptionInstance.addProperty (CIMProperty (CIMName ("Filter"),
+        filterPath, 0, PEGASUS_CLASSNAME_INDFILTER));
+    subscriptionInstance.addProperty (CIMProperty (CIMName ("Handler"),
+        handlerPath, 0, handlerClass));
+
+    return subscriptionInstance;
+}
+
+
+void test_isSubscription()
+{
+    CIMObjectPath filterPath = _buildFilterOrHandlerPath(
+        PEGASUS_CLASSNAME_INDFILTER,
+        "Filter1",
+        System::getFullyQualifiedHostName(),
+        PEGASUS_NAMESPACENAME_INTEROP);
+
+    CIMObjectPath handlerPath = _buildFilterOrHandlerPath(
+        PEGASUS_CLASSNAME_INDHANDLER_CIMXML,
+        "handler1",
+        System::getFullyQualifiedHostName(),
+        PEGASUS_NAMESPACENAME_INTEROP);        
+    
+    PEGASUS_TEST_ASSERT(IndicationService::_isSubscription(filterPath) == false);
+    PEGASUS_TEST_ASSERT(IndicationService::_isSubscription(handlerPath) == false);
+    /* TODO, Not finished Superceeded by tests in Subscription
+    CIMInstance subscription = _buildSubscriptionInstance(
+        filterPath,
+        PEGASUS_CLASSNAME_INDHANDLER_CIMXML,
+        handlerPath);
+
+    PEGASUS_TEST_ASSERT(IndicationService::_isSubscription(subscription) == true);
+    */
+}
+
+void test_setSystemNameInSubscription()
+{
+    CIMObjectPath filterPath = _buildFilterOrHandlerPath(
+        PEGASUS_CLASSNAME_INDFILTER,
+        "Filter1",
+        String::EMPTY,
+        PEGASUS_NAMESPACENAME_INTEROP);
+
+    CIMObjectPath handlerPath = _buildFilterOrHandlerPath(
+        PEGASUS_CLASSNAME_INDHANDLER_CIMXML,
+        "handler1",
+        String::EMPTY,
+        PEGASUS_NAMESPACENAME_INTEROP);
+        
+    CIMInstance subscription = _buildSubscriptionInstance(
+        filterPath,
+        PEGASUS_CLASSNAME_INDHANDLER_CIMXML,
+        handlerPath);
+
+    IndicationService::_setSystemNameInSubscription(
+        subscription, System::getFullyQualifiedHostName(),
+        true);
+
+    // TODO test result
+    
+}
+
 int main()
 {
     test_setSystemNameInHandlerFilter();
     test_setOrAddSystemNameInHandlerFilter();
     test_setSystemNameInHandlerFilterReference();
+    test_setSystemNameInSubscription();
 
     return 0;
 }